requirements_cut.txt
# Additional requirements for CUT model experiments with paired testing

# For evaluation metrics (SSIM, PSNR, NRMSE)
scikit-image>=0.19.0

# Core dependencies (should already be installed for UNSB)
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.0.0
h5py>=3.0.0

# For visualization and logging
wandb>=0.12.0
matplotlib>=3.3.0

# Utilities
packaging>=20.0
